split worker into a scheduler and worker
