+++ server/src/worker/worker.db-queue.ts
@@ -1,5 +1,5 @@
-import { Monitor, supportsGeoCheck } from "@/domain/monitors/monitor.types.js";
-import { IWorker, WorkerJobsPagination, WorkerJobSummary, WorkerMetrics } from "@/worker/worker.interface.js";
+import { supportsGeoCheck } from "@/domain/monitors/monitor.types.js";
+import { IQueueWorker } from "@/worker/worker.interface.js";
 import { type Job, type JobSeed, type JobType, jobId, LOCK_MS } from "@/domain/jobs/job.type.js";
 import { ILogger } from "@/utils/logger.js";
 import { IJobsRepository } from "@/domain/jobs/job.repository.interface.js";
@@ -14,6 +14,7 @@ import { IWorkerHelper } from "@/worker/worker.helper.js";
 import { IQueueWorkersRepository } from "@/domain/queue-workers/queue-worker.repository.interface.js";
 import { WORKER_TTL_SECONDS } from "@/domain/queue-workers/queue-worker.model.js";
 import { QueueMode } from "@/domain/app-settings/app-settings.type.js";
+import { JobScheduler } from "@/worker/worker.job-scheduler.js";
 const SERVICE_NAME = "JobQueue";
 const POLL_MS = 250; // base poll interval while a loop is actively claiming work
 const POLL_MAX_MS = 5000; // back off poll interval to 5s if no jobs
@@ -29,10 +30,8 @@ const CONCURRENCY: Record<JobType, number> = {
 	"cleanup-retention": 1,
 };
 
-export class DBQueueWorker implements IWorker {
+export class DBQueueWorker extends JobScheduler implements IQueueWorker {
 	static SERVICE_NAME = SERVICE_NAME;
-	private stopped = false;
-	private timers = new Map<JobType | "heartbeat", NodeJS.Timeout>();
 	private inFlight: Record<JobType, number> = {
 		check: 0,
 		"geo-check": 0,
@@ -41,33 +40,24 @@ export class DBQueueWorker implements IWorker {
 		"cleanup-retention": 0,
 	};
 
-	// Per job type poll intervals.  These back off on idle and reset on work or wake()
-	private pollMs: Record<JobType, number> = {
-		check: POLL_MS,
-		"geo-check": POLL_MS,
-		evaluate: POLL_MS,
-		"cleanup-orphaned": POLL_MS,
-		"cleanup-retention": POLL_MS,
-	};
-	private ticking: Partial<Record<JobType, boolean>> = {}; // Whether a loop is mid tick
-	private tickFns = new Map<JobType, () => void>(); // Store tick fns so we can run them immediately on wake
-
 	constructor(
 		private logger: ILogger,
-		private jobsRepository: IJobsRepository,
+		jobsRepository: IJobsRepository, // ← no modifier: forwarded to super
 		private monitorsRepository: IMonitorsRepository,
 		private checksRepository: IChecksRepository,
-		private checkService: ICheckService, // toStatusResponse + toLastEvaluatedAt
+		private checkService: ICheckService,
 		private checkProducer: ICheckProducer,
 		private checkEvaluator: ICheckEvaluator,
 		private geoCheckPipeline: ICheckPipeline,
 		private dispatcher: IReactorDispatcher,
 		private helper: IWorkerHelper,
-		private queueWorkersRepository: IQueueWorkersRepository,
+		queueWorkersRepository: IQueueWorkersRepository, // ← no modifier: forwarded
 		private queueMode: QueueMode,
 		private queuePrimaryProcesses: boolean,
-		private readonly workerId: string // same id MongoJobsRepository stamps into lockedBy
-	) {}
+		workerId: string
+	) {
+		super(jobsRepository, queueWorkersRepository, workerId);
+	}
 
 	get serviceName() {
 		return DBQueueWorker.SERVICE_NAME;
@@ -109,23 +99,6 @@ export class DBQueueWorker implements IWorker {
 		return instance;
 	}
 
-	// Helpers, builds job seeds. immediate=true runs on the next tick; otherwise jitter spreads the herd.
-	private toCheckJob = (monitor: Monitor, now: number, immediate = false): JobSeed => ({
-		id: jobId("check", monitor.id),
-		type: "check",
-		refId: monitor.id,
-		isActive: monitor.isActive,
-		nextScheduledAt: immediate ? now : now + Math.floor(Math.random() * monitor.interval), // Avoid herd
-		intervalMs: monitor.interval,
-	});
-
-	private toGeoCheckJob = (monitor: Monitor, now: number, immediate = false): JobSeed => ({
-		...this.toCheckJob(monitor, now, immediate),
-		id: jobId("geo-check", monitor.id),
-		type: "geo-check",
-		intervalMs: monitor.geoCheckInterval!,
-	});
-
 	private toCleanupJob = (type: "cleanup-orphaned" | "cleanup-retention", now: number): JobSeed => ({
 		id: jobId(type, null),
 		type,
@@ -133,20 +106,6 @@ export class DBQueueWorker implements IWorker {
 		isActive: true,
 		nextScheduledAt: now,
 		intervalMs: CLEANUP_INTERVAL_MS,
-	});
-
-	private toSummary = (job: Job): WorkerJobSummary => ({
-		monitorId: job.refId ?? job.id,
-		monitorType: job.type,
-		monitorInterval: job.intervalMs,
-		monitorActive: job.isActive,
-		lockedBy: job.lockedBy,
-		lockedUntil: job.lockedUntil,
-		nextScheduledAt: job.nextScheduledAt,
-		runCount: job.runCount,
-		failCount: job.failCount,
-		failReason: job.lastFailReason,
-		lastFinishedAt: job.lastFinishedAt,
 	});
 
 	// ********************
@@ -276,15 +235,6 @@ export class DBQueueWorker implements IWorker {
 
 	// Wake a worker loop after adding a job that's due immediately.
 	// Otherwise the loop could be delayed by POLL_MAX_MS if it is already fully backed off
-	wake = (type: JobType) => {
-		this.pollMs[type] = POLL_MS;
-		if (this.stopped || this.ticking[type]) return; // If it's not running or mid tick, do nothing
-		const tick = this.tickFns.get(type);
-		if (!tick) return; // Try to get a handle on the tick fn, if we can't we're done
-		const pending = this.timers.get(type);
-		if (pending) clearTimeout(pending); // Clear out scheduled tick if any
-		this.timers.set(type, setTimeout(tick, 0)); // Set immediate tick
-	};
 
 	// Seed queue
 	private reconcile = async () => {
@@ -332,85 +282,7 @@ export class DBQueueWorker implements IWorker {
 		return true;
 	};
 
-	addJob = async (_monitorId: string, monitor: Monitor) => {
-		const now = Date.now();
-		await this.jobsRepository.upsertJob(this.toCheckJob(monitor, now, true));
-		this.wake("check");
-		if (supportsGeoCheck(monitor.type) && monitor.geoCheckEnabled) {
-			await this.jobsRepository.upsertJob(this.toGeoCheckJob(monitor, now, true));
-			this.wake("geo-check");
-		}
-	};
-	deleteJob = async (monitor: Monitor) => {
-		await this.jobsRepository.deleteById(monitor.id);
-	};
-	pauseJob = async (monitor: Monitor) => {
-		await this.jobsRepository.setActiveById(monitor.id, false);
-	};
-	resumeJob = async (monitor: Monitor) => {
-		await this.jobsRepository.setActiveById(monitor.id, true); // a paused job's nextScheduledAt is now in the past, so it's due immediately
-		this.wake("check");
-		if (supportsGeoCheck(monitor.type) && monitor.geoCheckEnabled) this.wake("geo-check");
-	};
-	updateJob = async (monitor: Monitor) => {
-		await this.jobsRepository.updateScheduleById(monitor.id, "check", monitor.interval);
-		if (supportsGeoCheck(monitor.type) && monitor.geoCheckEnabled) {
-			await this.jobsRepository.upsertJob(this.toGeoCheckJob(monitor, Date.now(), true)); // a newly enabled geo check should run right away
-			this.wake("geo-check");
-		} else {
-			await this.jobsRepository.deleteByIdAndType(monitor.id, "geo-check");
-		}
-	};
-	shutdown = async () => {
-		this.stopped = true;
-		// Stop all timers
-		for (const timer of this.timers.values()) {
-			clearTimeout(timer);
-		}
-		this.timers.clear(); // Clear the map out
-		//  Remove workers from registry
-		await this.queueWorkersRepository.deleteById(this.workerId);
-	};
-	getMetrics = async (): Promise<WorkerMetrics> => {
-		const rows = await this.jobsRepository.findAll();
-		const workers = await this.queueWorkersRepository.findRecent(WORKER_STALE_MS);
-		const now = Date.now();
-		return rows.reduce<WorkerMetrics>(
-			(acc, job) => {
-				acc.jobs++;
-				acc.totalRuns += job.runCount;
-				acc.totalFailures += job.failCount;
-				if (job.lockedBy !== null && job.lockedUntil !== null && job.lockedUntil > now) acc.activeJobs++;
-				if (job.failCount > 0) {
-					acc.failingJobs++;
-					acc.jobsWithFailures.push({
-						monitorId: job.refId ?? job.id,
-						monitorUrl: null,
-						monitorType: job.type,
-						failedAt: job.lastFinishedAt,
-						failCount: job.failCount,
-						failReason: job.lastFailReason,
-					});
-				}
-				return acc;
-			},
-			{
-				jobs: 0,
-				activeJobs: 0,
-				failingJobs: 0,
-				jobsWithFailures: [],
-				totalRuns: 0,
-				totalFailures: 0,
-				workers,
-			}
-		);
-	};
-
-	getJobs = async (pagination: WorkerJobsPagination) => {
-		const { jobs, count } = await this.jobsRepository.findPage(pagination);
-		return { jobs: jobs.map(this.toSummary), count };
-	};
-	flushQueues = async () => {
+	override flushQueues = async () => {
 		await this.shutdown();
 		const ok = await this.init();
 		return { success: ok };
